# docs/swagger.yaml
basePath: /api/v1
definitions:
  apikey.CreateAPIKeyResponse:
    properties:
      api_key:
        $ref: '#/definitions/models.APIKeyResponse'
      message:
        type: string
      plain_key:
        description: Only returned on creation
        type: string
      success:
        type: boolean
    type: object
  apikey.DeleteAPIKeyResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  apikey.ListAPIKeysResponse:
    properties:
      api_keys:
        items:
          $ref: '#/definitions/models.APIKeyResponse'
        type: array
      message:
        type: string
      success:
        type: boolean
      total:
        type: integer
    type: object
  bucket.CreateBucketCommand:
    properties:
      auth_rule:
        $ref: '#/definitions/models.AuthRuleResponse'
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 63
        minLength: 3
        type: string
      owner_id:
        type: string
      settings:
        $ref: '#/definitions/models.BucketSettingsResponse'
    required:
    - name
    type: object
  bucket.CreateBucketResponse:
    properties:
      bucket:
        $ref: '#/definitions/models.BucketResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  bucket.DeleteBucketResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  bucket.GetBucketResponse:
    properties:
      bucket:
        $ref: '#/definitions/models.BucketResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  bucket.ListBucketsResponse:
    properties:
      buckets:
        items:
          $ref: '#/definitions/models.BucketResponse'
        type: array
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
  bucket.UpdateBucketCommand:
    properties:
      auth_rule:
        $ref: '#/definitions/models.AuthRuleResponse'
      bucket_id:
        type: string
      description:
        maxLength: 500
        type: string
      settings:
        $ref: '#/definitions/models.BucketSettingsResponse'
      user_id:
        type: string
    type: object
  bucket.UpdateBucketResponse:
    properties:
      bucket:
        $ref: '#/definitions/models.BucketResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.APIKeyPermission:
    properties:
      buckets:
        items:
          type: string
        type: array
      read:
        type: boolean
      sign_urls:
        type: boolean
      write:
        type: boolean
    type: object
...
  models.APIKeyResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      key_prefix:
        type: string
      last_used:
        type: string
      name:
        type: string
      permissions:
        $ref: '#/definitions/entities.APIKeyPermission'
      updated_at:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  models.AuthRuleResponse:
    properties:
      config:
        additionalProperties: true
        type: object
      enabled:
        type: boolean
      type:
        type: string
    type: object
  models.BucketResponse:
    properties:
      auth_rule:
        $ref: '#/definitions/models.AuthRuleResponse'
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      settings:
        $ref: '#/definitions/models.BucketSettingsResponse'
      stats:
        $ref: '#/definitions/models.BucketStatsResponse'
      updated_at:
        type: string
    type: object
  models.BucketSettingsResponse:
    properties:
      allow_overwrite:
        type: boolean
      allowed_extensions:
        items:
          type: string
        type: array
      allowed_mime_types:
        items:
          type: string
        type: array
      blocked_extensions:
        items:
          type: string
        type: array
      blocked_mime_types:
        items:
          type: string
        type: array
      encryption:
        type: boolean
      max_file_size:
        type: integer
      max_files_per_bucket:
        type: integer
      max_total_size:
        type: integer
      public_read:
        type: boolean
      require_content_type:
        type: boolean
      versioning:
        type: boolean
    type: object
  models.BucketStatsResponse:
    properties:
      last_access:
        type: string
      total_files:
        type: integer
      total_size:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@shbucket.local
    name: SHBucket Support
  description: SHBucket is a distributed object storage system similar to AWS S3
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SHBucket API
  version: 2.0.0
paths:
  /api-keys:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of API keys for the authenticated user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of API keys
          schema:
            $ref: '#/definitions/apikey.ListAPIKeysResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      description: Create a new API key for programmatic access
      parameters:
      - description: API key creation request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: API key created successfully
          schema:
            $ref: '#/definitions/apikey.CreateAPIKeyResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create API key
      tags:
      - api-keys
  /api-keys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an API key by ID
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: API key deleted successfully
          schema:
            $ref: '#/definitions/apikey.DeleteAPIKeyResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete API key
      tags:
      - api-keys
...
  /buckets:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of buckets for the authenticated user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of buckets
          schema:
            $ref: '#/definitions/bucket.ListBucketsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: List buckets
      tags:
      - buckets
    post:
      consumes:
      - application/json
      description: Create a new storage bucket with specified settings
      parameters:
      - description: Bucket creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/bucket.CreateBucketCommand'
      produces:
      - application/json
      responses:
        "201":
          description: Bucket created successfully
          schema:
            $ref: '#/definitions/bucket.CreateBucketResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Create new bucket
      tags:
      - buckets
...
  /buckets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a storage bucket by ID
      parameters:
      - description: Bucket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bucket deleted successfully
          schema:
            $ref: '#/definitions/bucket.DeleteBucketResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Delete bucket
      tags:
      - buckets
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific bucket
      parameters:
      - description: Bucket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bucket information
          schema:
            $ref: '#/definitions/bucket.GetBucketResponse'
        "400":
          description: Invalid bucket ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Bucket not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Get bucket by ID
      tags:
      - buckets
    put:
      consumes:
      - application/json
      description: Update bucket settings and metadata
      parameters:
      - description: Bucket ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/bucket.UpdateBucketCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Bucket updated successfully
          schema:
            $ref: '#/definitions/bucket.UpdateBucketResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Update bucket
      tags:
      - buckets
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for authentication
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"